--- extensions/internals/error_handler/constants.py
@@ -1,1 +1,0 @@
-import discord
